+++ src/sarsa/__about__.py
@@ -1,0 +1,5 @@
+# Author: Yuan Zhao <yuan.zhao@nih.gov>
+# Affiliation: Machine Learning Core, NIMH
+"""Package metadata."""
+
+__version__ = "0.4.0"